kernel: fix bug in FuncIS_ANTISYMMETRIC_DIGRAPH
The problem in line 704 was that it assumed incorrectly that the variable
nbs was a plain list, and hence tried to get the kth element of the list using
ADDR_OBJ(nbs)[k]. When nbs is range this returned nonsense, and
caused a seg fault. This commit fixes this issue.