kernel: fix bug in FuncDIGRAPH_TOPO_SORT
The problem was in line 864. The ADDR_OBJ macro was used on an
entry of the out-neighbours, without the guarantee that the entry
was a plain list. Thus, when given a range, this caused a seg
fault. This commit fixes this issue by using ELM_LIST(nbs, k)
instead of ADDR_OBJ(nbs)[k].