kernel: fix wrong size int bug in homomorphism finder
This caused an infinite loop in new_conditions when was called and the
first digraph had at least 256 vertices, the loop variable i was a UIntS
and compared to nr1 * nr1, which is 65536 or more with 256 vertices, and
so the loop condition was always true. This lead to an infinite loop
where more and more memory was assigned and nothing happened.