attr: improve symmetric closure digraph
This commit improves the permformance of DigraphSymmetricClosure, in the
particular case of digraphs without multiple edges. The previous method
involved creating the adjacency matrix of the digraph whose symmetric closure
we are finding, and if the number of vertices is too high, then this is
prohibitively expensive, both in memory and in time.