Remove two calls to RetypeBag from kernel extension
Motivation: RetypeBag is a very low-level part of the GAP kernel, and kernel
extensions should ideally restrict use of it to what's really required. In
many cases, it is far better to avoid it, and instead use higher-level APIs
such as ASS_LIST to edit lists, instead of relying on the unsafe low-level API
SET_ELM_PLIST combined with RetypeBag.

However, in the two cases at hand, calling RetypeBag simply is unnecessary: In
the first one, the code actually uses ASS_LIST and as such any retyping is
automatically performed by the GAP kernel for us.

In the second case, an empty plist is marked as such explicitly by retyping it
to T_PLIST_EMPTY. But that's a purely optional optimization.