display.tst: Change test that relies on xpdf not being installed.
Previously, the tests in display.tst try to call Splash with "xpdf" as the viewer, and we expect an error to be thrown.  This only works if xpdf is not installed, and if xpdf is installed then the tests fail.

This commit changes the test to a dummy viewer which will definitely not be installed.