cliques: some perf improvements
This commit includes some performance improvements for the clique
finding code.  The example in the newly added extreme/cliques.tst file
takes approx. 4s with the modified code, versus at least 3 minutes
previously (I never ran it to completion previously because it was
taking too long). If we do:

List(DigraphCliquesReps(D),
     x -> Orbit(AutomorphismGroup(D), x, OnSets));

for the same example, then this takes about 12s, doing the same with

List(DigraphCliquesReps(D),
     x -> Orb(AutomorphismGroup(D), x, OnSets));

instead runs out of memory repeatedly (and takes longer than 12s). Note
that the automorphism group is of size 16, and so the orbits are
relatively small in comparison to the number of cliques.