Rewrite AbstractHTML in HTML format
This is not ideal because we should probably reuse the abstract
definition as provided to the Autodoc package.

--------------------------------

AbstractHTML is the text displayed on the package overview web page of
GAP.

AutoDoc Abstract is the text displayed on the title (home) page of digraphs.

The Autodoc package uses the GAPDoc specification for input [1],
whereas AbstractHTML, provided by GAP itself [2],
requires the string in HTML.

There may be a way to convert one to the other,
e.g. convert AutoDoc.TitlePage.Abstract from GAPDoc-formatted
into HTML-formatted string.
Possibly the converter and parser functions
ParseTreeXMLString and GAPDoc2HTML provided by GAPDoc package could be
useful.

Another thing might be AutoDoc entities, as detailed in Chapter 4 of
their manual.

[1] -- AutoDOC Reference Manual
[2] -- GAP Reference Manual, Chapter 76, Section 76.3-15: "The
PackageInfo.g File"