Expand named entity GAPDoc calls in Abstract(HTML)
Derived from XML, GAPDoc has macro substitution in the form of named
entities. Calling a macro, e.g. &Digraphs;, will expand the html
defined for the given named entity.
The intended behaviour of the calls is exhibited in Chapter 0 of
the project manual [1].
More information about named entities can be found in Chapter 2.1-9
of the GAPDoc documentation manual [2].

AutoDoc doesn't seem to expand these macros. This results in the
raw macro call being displayed in the places controlled by AutoDoc,
namely the project homepage [3] and Digraphs' entry on the
list of GAP packages [4].

Hence to prevent unnecessarily defining the Abstract twice,
with different styling for GAPDoc and AutoDoc,
instead expand the macro in-place.

Note this means any future changes to the macro (named entity)
will not propagate here.

[1] -- https://digraphs.github.io/Digraphs/doc/chap0_mj.html
[2] -- https://docs.gap-system.org/pkg/gapdoc/doc/chap2.html
[3] -- https://digraphs.github.io/Digraphs/
[4] -- https://www.gap-system.org/packages/