Add doc/index.html to redirect <homepage>/doc to the title page of the manual
This means that if a user requests https://digraphs.github.io/Digraphs/doc/
(which I think is a natural-enough thing to want to do), then instead of
getting a 404 error, they are instead automatically redirected to the
title page of the manual, which is presumably what they wanted, anyway.

Unless and until https://github.com/gap-system/GitHubPagesForGAP/pull/25
is merged, or https://github.com/gap-actions/update-gh-pages/issues/10
is addressed in a way that lets us keeps this file purely in our gh-pages
branch, then I think the best way to manage this file is to keep it in
the repository here.

Whenever the ReleaseTools script/release action is run, the doc/
directory of the gh-pages branch is deleted, and so if we only kept this
file in the gh-pages branch, it would get deleted each time we make a
new release. However, the doc/ directory of the gh-pages branch is
re-populated with (amongst other things) the html files of the doc/
directory of this repository. In particular, by having this file in the
doc/ directory of this repo, this means that the file is retained in the
gh-pages branch, even across different releases of the package and runs
of the ReleaseTools script/release action.